refactor: the stack parts separations
