+++ VirtualFinland.UsersAPI.Deployment/VirtualFinland.UsersAPI.Deployment.csproj
@@ -8,10 +8,10 @@
 
   <ItemGroup>
     <PackageReference Include="Pulumi" Version="3.*" />
-    <PackageReference Include="Pulumi.Aws" Version="5.16.2" />
+    <PackageReference Include="Pulumi.Aws" Version="5.40.0" />
     <PackageReference Include="Pulumi.Command" Version="4.5.0" />
     <PackageReference Include="Pulumi.Random" Version="4.8.2" />
-    <PackageReference Include="Pulumi.Awsx" Version="1.0.0-beta.11" />
+    <PackageReference Include="Pulumi.Awsx" Version="1.0.2" />
   </ItemGroup>
 
 </Project>
